Update patch set 1
Patch Set 1:

(sorry for the delay on this. hopefully today i'll be able to flip the switch so mips is using the uapi headers, but then we'll probably have to revisit this. is this fixed in upstream strace? it's high time we tried to get back in sync...)

Patch-set: 1
Label: Code-Review=0