Update patch set 2
Patch Set 2:

stat of x86_64 is not same as stat64 of i386. So we can't define the stat64 to stat for x86_64.

And strace can handle this issue for x86_64. Please check the line 1029 of file.c

I checked the upstream strace on PC side. Its config.h has HAVE_STAT64 undefined and let line 1029 in file.c handle it.

Patch-set: 2