Update patch set 2
Patch Set 2:

Tested the patch http://sourceforge.net/p/strace/mailman/message/32093102/.
It can't fix the issue on x86_64.

I suppose the reason is the difference of stat for x86_64 and stat64 for i386. So the line 1029 of file.c introduces a structure. This structure has same memory layout on x86_64 as stat64 on i386.

Patch-set: 2