Update patch set 4
Patch Set 4:

actually, should we build an strace32 too? it might be a good way to ensure that we don't accidentally break the 32-bit architectures as we transition to 64-bit probably being the default that everyone works with. we could make strace32 optional so it doesn't get built/take up space unless you're actually working on strace itself and build in this directory.

Patch-set: 4