\#undef various <sys/stat.h> #defines.
This is part of a set of changes that get us a POSIX struct stat.
To maintain compatibility with existing Android source, I need to pollute
the namespace with a few extra macros. We need to decide whether we want
to keep these #defines long term (in which case we should upstream this
change) or not (in which case we should rever this change as soon as
we've moved all the platform code to the POSIX names --- something I
think we should do regardless).

Bug: 18298106
Change-Id: Ife4f574e5d5d86add5e951ea668456025181ca1f