Create change
Uploaded patch set 1.

Patch-set: 1
Change-id: Ife4f574e5d5d86add5e951ea668456025181ca1f
Subject: This is part of a set of changes that get us a POSIX struct stat. To maintain compatibility with existing Android source, I need to by this. We need to decide whether we want to keep these #defines long term (in which case we should upstream this change) or not (in which case we should rever this change as soon as we've moved all the platform code to the POSIX names --- something I think we should do regardless).
Branch: refs/heads/master
Commit: b9904dc58693ec4a9b6fca95221b54b0af64db73