Revert "Fix our pre-GPL strace to build as C23."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e32a6cbf-8330-4b7a-94d8-1264f188c16c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2058993e01ac18e5187202e0228f166cb3bda2a1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com